requirements.txt
streamlit
python-dotenv
google-genai
fastapi
uvicorn
requests
